# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample swagger for Fiber",
        "title": "Fiber Example API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "email": "fiber@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/product/category/add": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "add category",
                "tags": [
                    "Category"
                ],
                "summary": "add category",
                "parameters": [
                    {
                        "description": "prev - cur",
                        "name": "Cate",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/product.Category"
                        }
                    },
                    {
                        "type": "string",
                        "description": "Authorization",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
        "/product/category/refresh": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "remake category_map in server",
                "tags": [
                    "Category"
                ],
                "summary": "refresh category",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authorization",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
        "/test": {
            "get": {
                "description": "Test get list",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Test"
                ],
                "summary": "Test List",
                "responses": {}
            }
        },
        "/user/login": {
            "post": {
                "description": "Check ID and Password\nPassword secure by bcrypt",
                "tags": [
                    "User"
                ],
                "summary": "Login",
                "parameters": [
                    {
                        "description": "ID and Password",
                        "name": "ID\u0026PWD",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/utils.LoginInfo"
                        }
                    }
                ],
                "responses": {}
            }
        },
        "/user/signup": {
            "post": {
                "description": "Post user info for sign up into DB",
                "tags": [
                    "User"
                ],
                "summary": "Sign up",
                "parameters": [
                    {
                        "description": "User infomation",
                        "name": "INFO",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/user.User"
                        }
                    }
                ],
                "responses": {}
            }
        },
        "/user/{id}/info": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Get User Info from DB",
                "tags": [
                    "User"
                ],
                "summary": "Get User Info",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Authorization",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
        "/vender/login": {
            "post": {
                "description": "Check ID and Password\nPassword secure by bcrypt",
                "tags": [
                    "Vender"
                ],
                "summary": "Login",
                "parameters": [
                    {
                        "description": "ID and Password",
                        "name": "ID\u0026PWD",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/utils.LoginInfo"
                        }
                    }
                ],
                "responses": {}
            }
        },
        "/vender/signup": {
            "post": {
                "description": "vender sign up",
                "tags": [
                    "Vender"
                ],
                "summary": "vender sign up",
                "parameters": [
                    {
                        "description": "vender info",
                        "name": "INFO",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/vender.Vender"
                        }
                    }
                ],
                "responses": {}
            }
        },
        "/vender/{id}/info": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "get vender info",
                "tags": [
                    "Vender"
                ],
                "summary": "get\tvender info",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authorization",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {}
            }
        }
    },
    "definitions": {
        "product.Category": {
            "type": "object",
            "properties": {
                "current cate": {
                    "type": "string"
                },
                "previous cate": {
                    "type": "string"
                }
            }
        },
        "user.User": {
            "type": "object",
            "properties": {
                "addr": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "phone_number": {
                    "type": "string"
                }
            }
        },
        "utils.LoginInfo": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
        "vender.Vender": {
            "type": "object",
            "properties": {
                "addr": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "phone_number": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "description": "OAuth protects our entity endpoints",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}